requirements.txt
flask==1.1.2
requests==2.25.1
sqlalchemy==1.3.23
flask-login==0.5.0
flask-sqlalchemy==2.4.4
mysql-connector-python==8.0.23
wget
